requirements.txt
-e git+https://github.com/marcoancona/DeepExplain.git@6ac43e729dc8db6ce05ca241dfeb35f1c3eb5b0b#egg=deepexplain
innvestigate==1.0.8
deeplift==0.6.8.1
keras-vis==0.4.1
shap==0.26.0
efficientnet==0.0.2
future==0.17.1
h5py==2.7.1
imageio==2.4.1
imgaug==0.2.6
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keras-resnet==0.1.0
matplotlib==3.0.3
mysqlclient==1.4.0
numpy==1.15.4
opencv-python==3.4.2.17
pandas==0.24.1
Pillow==5.4.1
py==1.7.0
pydicom==2.0.0
scikit-image==0.14.0
scikit-learn==0.20.2
scipy==1.2.1
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
